requirements.txt
numpy
onnx
onnxruntime-gpu
opencv-python
torch
torchvision
tqdm
